Title: 컴퓨터와 함께하는 숫자 게임 2, Time: 12ms, Memory: 0MB, Status: Passed - CodeTree

+++ 241023/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 unsigned long long m;
 
-unsigned long long solve(int x)
+unsigned long long solve(unsigned long long x)
 {
     unsigned long long left = 1, right = m;
     unsigned long long cnt = 0;
